Move all classes into 'Coffee' package
