+++ src/main/java/Coffee/ByeCommand.java
@@ -1,3 +1,5 @@
+package Coffee;
+
 public class ByeCommand extends Command {
 
     @Override
