+++ src/main/java/Coffee/Command.java
@@ -1,3 +1,5 @@
+package Coffee;
+
 public abstract class Command {
 
     public abstract void execute(TaskList tasks, Ui ui, Storage storage);
