+++ src/main/java/Coffee/DeleteCommand.java
@@ -1,3 +1,5 @@
+package Coffee;
+
 public class DeleteCommand extends Command {
 
     private final int index;
