+++ src/main/java/Coffee/Event.java
@@ -1,3 +1,5 @@
+package Coffee;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
