+++ src/main/java/Coffee/EventCommand.java
@@ -1,3 +1,5 @@
+package Coffee;
+
 public class EventCommand extends Command {
     private final String description;
     private final String from;
