+++ src/main/java/Coffee/LineBreak.java
@@ -1,3 +1,5 @@
+package Coffee;
+
 public class LineBreak {
 
     private static String lineBreak =
