+++ src/main/java/Coffee/Storage.java
@@ -1,3 +1,5 @@
+package Coffee;
+
 import java.io.BufferedWriter;
 import java.io.FileWriter;
 import java.io.IOException;
