+++ src/main/java/Coffee/Task.java
@@ -1,3 +1,5 @@
+package Coffee;
+
 public class Task {
     protected String description;
     protected boolean isDone;
